requirements.txt
pip>=23.0
setuptools>=65.0
wheel>=0.38

python-docx
streamlit==1.35.0
sentence-transformers==2.2.2
transformers==4.28.1

torch>=1.9.0
numpy>=1.21
pandas>=1.5
scikit-learn>=1.2
PyPDF2==3.0.1
google-api-python-client==2.127.0
google-auth-httplib2
google-auth-oauthlib==1.2.0
google-auth==2.29.0
huggingface_hub
tabulate


